pyproject.toml
# Note: Ths file is generated by the cookiecutter template and modify with poetry commands. It's not neccessary to
# edit it manually (advanced).
# The pyproject.toml file with Poetry: https://python-poetry.org/docs/pyproject/
# Fast Reference: https://py-pkgs.org/03-how-to-package-a-python#summary-and-next-steps

[tool.poetry]
name = "dupchi"
version = "0.2.0"
description = "Denoise, Upscale and Predict Cancer from Histopathological Images"
authors = ["Javier Yago Córcoles <javier.yago14@gmail.com>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
# Core Packages:  poetry add <package>  (to remove: poetry remove <package>)
# Common issue: In case of incompatibility error after 'poetry add <package>', just set the --python property to the one
# suggested at the end of the error message. e.g.:
# output message: ... a possible solution would be to set the `python` property to ">=3.11,<3.12"
# solution: $ poetry add pandas-profiling --python ">=3.11,<3.12"

python = "^3.10"
pandas = ">=2.1"
pyarrow = ">=13"
py-cpuinfo = ">=9.0.0"
python-dotenv = ">=1.0.0"
numpy = '>=1.26.4'
tensorflow = '>=2.16.1'
matplotlib = '>=3.8.3'
pillow = '>=10.2.0'
torch = '>=2.2.2'
tqdm = '>=4.66.2'
albumentations = '==1.4.2'
torchvision = '>=0.17.2'
super-image = '>=0.1.7'
gradio = '>=4.24.0'
google-cloud-storage = '>=2.16.0'
opencv-python = '>=4.9.0.80'

# sentence-transformers = ">=2.2.2"

# LLM
# langchain = ">=0.1.0"
# openai = ">=1.0.0"
# tiktoken = ">=0.5.1"

# Retrieval / Database
# faiss-cpu = ">=1.7.4"

# google-cloud-aiplatform = ">=1.38.1"


[tool.poetry.extras]
# Advanced: recommended for large Building Blocks only
# Detailed Dependence Groups with Poetry: https://python-poetry.org/docs/managing-dependencies/#dependency-groups
# e.g.: $ poetry add tensorflow --extras "deep_learning" --optional

[tool.poetry.group.dev.dependencies]
# Edition Only Packages:  poetry add <package> --group dev (to remove: poetry remove <package>)
# They won't be on the final package (pip install <my-package>), but will be installed when developing the package
notebook = ">=7"
jupyter = ">=1.0.0"
jupyter-contrib-nbextensions = ">=0.7.0"

# Plottting
# plotly = ">=5.17.0"

# Front Lite
# streamlit = ">=1.27.2"

# API
# fastapi = ">=0.100.0"
# uvicorn = ">=0.23.1"


pytest-cov = ">=4.1"
pre-commit = ">=3.4.0"
mypy = ">=1.5"
flake8 = ">=6.1.0"
pylint = ">=2.17"
isort = ">=5.12"
black = ">=23.9"
pyupgrade = ">=3.12"
snakeviz = ">=2.2"





# version_variable = "pyproject.toml:version" # version location
# branch = "main"                             # branch to make releases of
# changelog_file = "CHANGELOG.md"             # changelog file
# build_command = "poetry build"              # build dists
# dist_path = "dist/"                         # where to put dists
# upload_to_release = true                    # auto-create GitHub release
# upload_to_pypi = false                      # don't auto-upload to PyPI
# remove_dist = false                         # don't remove dists
# patch_without_tag = true                    # patch release by default
# Poetry api for building and publishing (do not change)
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
